basic popover works
