+++ app/components/popover-on-element.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-tooltips/components/popover-on-element';
